requirements.txt
timeitsklearn.model_selection sklearn.pipeline imblearn.pipeline sklearn.preprocessing sklearn.feature_selection sklearn.linear_model sklearn.naive_bayes sklearn.tree sklearn.svm sklearn.model_selection sklearn.preprocessing sklearn.ensemble sklearn.preprocessing joblibcategory_encoders sklearn.compose sklearn.neighbors imblearn.over_sampling sklearn.impute imblearn.under_sampling xgboost sklearn.metrics sklearn.model_selectionseabornnumpyplotly.expresspandasmatplotlib.pyplotColumnTransformerOneHotEncoderStandardScalerDecisionTreeClassifierModelBinaryEncoder